Adding the two CFX messages:
- "RecipeChanged"
A recipe describes, starting with the panel as root object, all components to inspect. It does not (yet) contain any inspection results.
The name is inspired from SECS/GEM "recipe". Synonyms are "inspection program" (in the domain / field of AOI/AXI) or "data model" (in Viscom vVision).

- "DefectsDetected"
List of only the defect components. Always refers to a previously sent recipe (i.e. sent in a RecipeChanged message).