Don't deadlock when calling ExecuteRequest.
The application will almost always deadlock when blocking on "async" functions. Commit c48a4a94ba7b95921834444744c4ba5963b53fe1
(by me) introduced ExecuteRequestAsync and made ExecuteRequest a wrapper around ExecuteRequestAsync. This results in a deadlock when
the synchronous method is called. Instead of blocking on the async method, run the request on a thread and wait for it.